central_server/processing_service/requirements.txt
pika
pydantic
google.genai
polars
numpy
sentence-transformers
python-dotenv
tzdata
SQLAlchemy
psycopg2-binary # For PostgreSQL connectivity
pgvector
pyarrow
pandas
tabulate
asyncpg
